+++ geoprox-core/src/lib.rs
@@ -1,6 +1,8 @@
-pub mod cache;
 mod metric;
+
+pub mod cache;
 pub mod models;
+pub mod serde;
 pub mod shard;
 
 pub use geohash;